added contract to EF model and fixed build issues

+++ RouteCoin/Program.cs
@@ -314,7 +314,7 @@ namespace RouteCoin
 
         private static bool IsBaseStationClose()
         {
-            return Math.Sqrt(Math.Pow(Math.Abs(node.PositionX - baseStationNode.PositionX), 2) + Math.Pow(Math.Abs(node.PositionY - baseStationNode.PositionY), 2)) < CoverageArea;
+            return Math.Sqrt(Math.Pow(Math.Abs((int)node.PositionX - (int)baseStationNode.PositionX), 2) + Math.Pow(Math.Abs((int)node.PositionY - (int)baseStationNode.PositionY), 2)) < CoverageArea;
         }
 
     }
